# components/admin-api-spec/openapi/essentials-admin-api.yaml
openapi: 3.0.3
info:
  title: Essentials Admin API
  description: |-
    Proposed HTTP admin/monitoring contract for the Essentials components (fenced locks, durable queues, scheduler, event store, CDC and PostgreSQL statistics).

    This contract is generated code-first from the in-process `*Api` SPI interfaces and is the single source of truth for both the generated Java client and the HTTP adapter that serves it (`spring-boot-starter-admin-api`).

    This document deliberately prescribes no authentication scheme. Authorization is role-based: each operation lists the roles that satisfy it under the `x-required-roles` vendor extension, and the `ESSENTIALS_ADMIN` role satisfies every operation. Those decisions are made by the hosting application's `EssentialsSecurityProvider` implementation, against the caller reported by its `EssentialsAuthenticatedUser` implementation. How the caller was authenticated is the host's business.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /api/essentials/admin/v1
  description: Relative to the host that mounts the Essentials admin API
...
paths:
...
  /durable-queues:
    get:
      tags:
      - durable-queues
      summary: List the names of all accessible durable queues.
      operationId: getQueueNames
      responses:
        "200":
          description: 200 — The accessible queue names.
          content:
            application/json:
              schema:
                uniqueItems: true
                type: array
                items:
                  type: string
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
  /durable-queues/messages/{queueEntryId}:
    get:
      tags:
      - durable-queues
      summary: Get a single queued message by its entry id.
      operationId: getQueuedMessage
      parameters:
      - name: queueEntryId
        in: path
        description: The queue entry id.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — The queued message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiQueuedMessage"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No value exists for the given identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
    delete:
      tags:
      - durable-queues
      summary: Delete a message from its queue.
      operationId: deleteMessage
      parameters:
      - name: queueEntryId
        in: path
        description: The queue entry id.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — Whether the message was deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_writer
      - essentials_admin
  /durable-queues/messages/{queueEntryId}/queue-name:
    get:
      tags:
      - durable-queues
      summary: Resolve the queue name owning a given message entry id.
      operationId: getQueueNameFor
      parameters:
      - name: queueEntryId
        in: path
        description: The queue entry id.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — The resolved queue name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueNameResult"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No value exists for the given identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
  /durable-queues/messages/{queueEntryId}/resurrect:
    post:
      tags:
      - durable-queues
      summary: "Resurrect a dead-letter message, re-queuing it after an optional delay."
      operationId: resurrectDeadLetterMessage
      parameters:
      - name: queueEntryId
        in: path
        description: The dead-letter message entry id.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResurrectDeadLetterMessageRequest"
        required: true
      responses:
        "200":
          description: 200 — The resurrected message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiQueuedMessage"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No value exists for the given identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_writer
      - essentials_admin
  /durable-queues/messages/{queueEntryId}/mark-as-dead-letter:
    post:
      tags:
      - durable-queues
      summary: Mark a queued message as a dead-letter message.
      operationId: markAsDeadLetterMessage
      parameters:
      - name: queueEntryId
        in: path
        description: The queue entry id.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — The updated message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiQueuedMessage"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No value exists for the given identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_writer
      - essentials_admin
  /durable-queues/queues/{queueName}/messages/count:
    get:
      tags:
      - durable-queues
      summary: Count messages queued for a queue.
      operationId: getTotalMessagesQueuedFor
      parameters:
      - name: queueName
        in: path
        description: The queue name.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — The total count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountResult"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
  /durable-queues/queues/{queueName}/dead-letter-messages/count:
    get:
      tags:
      - durable-queues
      summary: Count dead-letter messages queued for a queue.
      operationId: getTotalDeadLetterMessagesQueuedFor
      parameters:
      - name: queueName
        in: path
        description: The queue name.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — The total count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountResult"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
  /durable-queues/queues/{queueName}/messages:
    get:
      tags:
      - durable-queues
      summary: "List queued messages for a queue (paginated, sortable)."
      operationId: getQueuedMessages
      parameters:
      - name: queueName
        in: path
        description: The queue name.
        required: true
        schema:
          type: string
      - name: sortOrder
        in: query
        description: Sort order by queue entry id.
        required: false
        schema:
          type: string
          default: ASC
          enum:
          - ASC
          - DESC
      - name: startIndex
        in: query
        description: Zero-based index of the first row to return.
        required: false
        schema:
          type: integer
          format: int64
          default: 0
      - name: pageSize
        in: query
        description: Maximum number of rows to return.
        required: false
        schema:
          type: integer
          format: int64
          default: 100
      responses:
        "200":
          description: 200 — The matching items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiQueuedMessage"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
    delete:
      tags:
      - durable-queues
      summary: Purge all messages (including dead-letters) from a queue.
      operationId: purgeQueue
      parameters:
      - name: queueName
        in: path
        description: The queue name.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — Number of messages purged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurgeResult"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_writer
      - essentials_admin
  /durable-queues/queues/{queueName}/dead-letter-messages:
    get:
      tags:
      - durable-queues
      summary: "List dead-letter messages for a queue (paginated, sortable)."
      operationId: getDeadLetterMessages
      parameters:
      - name: queueName
        in: path
        description: The queue name.
        required: true
        schema:
          type: string
      - name: sortOrder
        in: query
        description: Sort order by queue entry id.
        required: false
        schema:
          type: string
          default: ASC
          enum:
          - ASC
          - DESC
      - name: startIndex
        in: query
        description: Zero-based index of the first row to return.
        required: false
        schema:
          type: integer
          format: int64
          default: 0
      - name: pageSize
        in: query
        description: Maximum number of rows to return.
        required: false
        schema:
          type: integer
          format: int64
          default: 100
      responses:
        "200":
          description: 200 — The matching items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiQueuedMessage"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
  /durable-queues/queues/{queueName}/statistics:
    get:
      tags:
      - durable-queues
      summary: Get delivery statistics for a queue.
      operationId: getQueuedStatistics
      parameters:
      - name: queueName
        in: path
        description: The queue name.
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 200 — The queue statistics.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiQueuedStatistics"
        "400":
          description: "Malformed or out-of-range parameter, or invalid request body."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Caller lacks one of the required roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No value exists for the given identifier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-required-roles:
      - essentials_queue_reader
      - essentials_admin
...
components:
  schemas:
...
    ApiQueuedMessage:
      required:
      - id
      - isBeingDelivered
      - isDeadLetterMessage
      - queueName
      - redeliveryAttempts
      - totalDeliveryAttempts
      type: object
      properties:
        id:
          type: string
        queueName:
          type: string
        payload:
          type: string
          description: The raw message payload. Null unless the caller holds the QUEUE_PAYLOAD_READER
            or ESSENTIALS_ADMIN role.
          nullable: true
        addedTimestamp:
          type: string
          format: date-time
        nextDeliveryTimestamp:
          type: string
          format: date-time
        deliveryTimestamp:
          type: string
          format: date-time
        lastDeliveryError:
          type: string
        totalDeliveryAttempts:
          type: integer
          format: int32
        redeliveryAttempts:
          type: integer
          format: int32
        isDeadLetterMessage:
          type: boolean
        isBeingDelivered:
          type: boolean
    ApiQueuedStatistics:
      required:
      - avgDeliveryLatencyMs
      - queueName
      - totalMessagesDelivered
      type: object
      properties:
        queueName:
          type: string
        fromTimestamp:
          type: string
          format: date-time
        totalMessagesDelivered:
          type: integer
          format: int64
        avgDeliveryLatencyMs:
          type: integer
          format: int32
        firstDelivery:
          type: string
          format: date-time
        lastDelivery:
          type: string
          format: date-time
...
    CountResult:
      required:
      - total
      type: object
      properties:
        total:
          type: integer
          format: int64
      description: A total count.
    DeleteResult:
      required:
      - deleted
      type: object
      properties:
        deleted:
          type: boolean
      description: Outcome of a message deletion.
    Error:
      required:
      - error
      - status
      type: object
      properties:
        status:
          type: integer
          description: The HTTP status code.
          format: int32
        error:
          type: string
          description: "Short, stable reason phrase for the status."
        message:
          type: string
          description: Human-readable detail. May be omitted where detail would leak
            internals.
          nullable: true
      description: Error response.
...
    PurgeResult:
      required:
      - purgedCount
      type: object
      properties:
        purgedCount:
          type: integer
          format: int32
      description: Number of messages removed by a purge.
    QueueNameResult:
      required:
      - queueName
      type: object
      properties:
        queueName:
          type: string
      description: A resolved queue name.
...
    ResurrectDeadLetterMessageRequest:
      required:
      - deliveryDelay
      type: object
      properties:
        deliveryDelay:
          type: string
          description: ISO-8601 delay before re-delivery; PT0S means immediate.
          format: duration
          example: PT0S
      description: Re-queue parameters for resurrecting a dead-letter message.